+++ components/app/explore/DatasetWidget.js
@@ -7,7 +7,7 @@ import { Autobind } from 'es-decorators';
 // Redux
 import { connect } from 'react-redux';
 import { Link } from 'routes';
-import { toggleLayerGroup } from 'redactions/explore';
+import { toggleLayerGroup, addFavoriteDataset, removeFavoriteDataset } from 'redactions/explore';
 import { toggleTooltip } from 'redactions/tooltip';
 
 // Components
@@ -17,9 +17,14 @@ import DatasetWidgetChart from 'components/app/explore/DatasetWidgetChart';
 import DatasetLayerChart from 'components/app/explore/DatasetLayerChart';
 import DatasetPlaceholderChart from 'components/app/explore/DatasetPlaceholderChart';
 import DatasetTagsTooltip from 'components/app/explore/DatasetTagsTooltip';
+import Spinner from 'components/ui/Spinner';
+
+// Services
+import GraphService from 'services/GraphService';
+import UserService from 'services/UserService';
 
 // Utils
-import { findTagInSelectorTree } from 'utils/explore/TreeUtil';
+import { TAGS_BLACKLIST } from 'utils/graph/TagsUtil';
 
 class DatasetWidget extends React.Component {
   /**
@@ -51,6 +56,23 @@ class DatasetWidget extends React.Component {
       x: window.scrollX + e.clientX,
       y: window.scrollY + e.clientY
     };
+  }
+
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      inferredTags: [],
+      loading: false
+    };
+
+    // Services
+    this.graphService = new GraphService({ apiURL: process.env.WRI_API_URL });
+    this.userService = new UserService({ apiURL: process.env.WRI_API_URL });
+  }
+
+  componentDidMount() {
+    this.loadInferredTags();
   }
 
   /**
@@ -102,11 +124,26 @@ class DatasetWidget extends React.Component {
     return null;
   }
 
-  getTags() {
-    const { dataset, topicsTree } = this.props;
+  /**
+  * Load inferred tags
+  */
+  loadInferredTags() {
+    const { dataset } = this.props;
     const vocabulary = dataset.attributes.vocabulary && dataset.attributes.vocabulary[0];
     const tags = vocabulary.attributes.tags;
-    return tags.filter(tag => findTagInSelectorTree(topicsTree, tag));
+    this.graphService.getInferredTags(tags)
+      .then((response) => {
+        this.setState({
+          inferredTags: response.filter(tag => tag.labels
+            .find(type => type === 'TOPIC' || type === 'GEOGRAPHY') &&
+            !TAGS_BLACKLIST.includes(tag.id)
+          )
+        });
+      })
+      .catch((err) => {
+        this.setState({ inferredTags: [] });
+        console.error(err);
+      });
   }
 
   /**
@@ -120,7 +157,7 @@ class DatasetWidget extends React.Component {
 
   @Autobind
   handleTagsClick(event) {
-    const tags = this.getTags();
+    const { inferredTags } = this.state;
 
     const position = DatasetWidget.getClickPosition(event);
     this.props.toggleTooltip(true, {
@@ -130,28 +167,64 @@ class DatasetWidget extends React.Component {
       childrenProps: {
         toggleTooltip: this.props.toggleTooltip,
         onTagClick: this.handleTagClick,
-        tags
+        tags: inferredTags
       }
     });
   }
 
   @Autobind
   handleTagClick(event) {
-    const tagName = event.target.textContent;
+    const tagName = event.target.getAttribute('id');
     this.props.onTagSelected(tagName);
   }
 
+  @Autobind
+  handleFavoriteButtonClick() {
+    const { favorite, dataset, user } = this.props;
+    this.setState({ loading: true });
+
+    if (!favorite) {
+      this.userService.createFavouriteDataset(dataset.id, user.token)
+        .then((response) => {
+          this.props.addFavoriteDataset(response.data, user.token);
+          this.setState({ loading: false });
+        })
+        .catch((err) => {
+          this.setState({ loading: false });
+          console.error(err);
+        });
+    } else {
+      this.userService.deleteFavourite(favorite.id, user.token)
+        .then((response) => {
+          this.props.onFavoriteRemoved(favorite);
+          this.props.removeFavoriteDataset(response.data);
+          this.setState({ loading: false });
+        })
+        .catch((err) => {
+          this.setState({ loading: false });
+          console.error(err);
+        });
+    }
+  }
+
   render() {
-    const { widget, layer, mode } = this.props;
+    const { widget, layer, mode, showActions, favorite, user } = this.props;
+    const { inferredTags, loading } = this.state;
     const dataset = this.props.dataset.attributes;
     const metadata = dataset.metadata && dataset.metadata[0];
-    const vocabulary = dataset.vocabulary && dataset.vocabulary[0];
-    const { showActions } = this.props;
     const gridMode = (mode === 'grid');
     const element = this.getWidgetOrLayer();
+    const starIconName = favorite ? 'icon-star-full' : 'icon-star-empty';
+
+    const starIconClass = classnames({
+      '-small': true,
+      '-filled': favorite,
+      '-empty': !favorite
+    });
 
     return (
       <div className={`c-dataset-list-item -${mode}`}>
+        <Spinner isLoading={loading} className="-small -light" />
 
         {/* If it has widget we want to renderize the default widget one */}
         {widget && gridMode &&
@@ -193,7 +266,7 @@ class DatasetWidget extends React.Component {
                 </Link>
               </h4>
               {/* Dataset tags link */}
-              {vocabulary && this.getTags().length &&
+              {inferredTags && inferredTags.length && inferredTags.length > 0 &&
                 <div
                   className="tags-button"
                   onClick={this.handleTagsClick}
@@ -204,6 +277,21 @@ class DatasetWidget extends React.Component {
                   <Icon
                     name="icon-item-category"
                     className="-small"
+                  />
+                </div>
+              }
+              {/* Favorite dataset icon */}
+              {user && user.id &&
+                <div
+                  className="favorite-button"
+                  onClick={this.handleFavoriteButtonClick}
+                  title="Favorite dataset"
+                  role="button"
+                  tabIndex={-1}
+                >
+                  <Icon
+                    name={starIconName}
+                    className={starIconClass}
                   />
                 </div>
               }
@@ -245,9 +333,11 @@ DatasetWidget.propTypes = {
   layer: PropTypes.object,
   mode: PropTypes.string,
   showActions: PropTypes.bool,
+  favorite: PropTypes.object,
 
   // Callbacks
   onTagSelected: PropTypes.func,
+  onFavoriteRemoved: PropTypes.func,
 
   // STORE
   // Topics tree used in the Explore selector
@@ -256,17 +346,23 @@ DatasetWidget.propTypes = {
   isLayerGroupAdded: PropTypes.func.isRequired,
   // Add or remove a layer group from the map
   toggleLayerGroup: PropTypes.func.isRequired,
-  toggleTooltip: PropTypes.func.isRequired
+  toggleTooltip: PropTypes.func.isRequired,
+  removeFavoriteDataset: PropTypes.func.isRequired,
+  addFavoriteDataset: PropTypes.func.isRequired,
+  user: PropTypes.object.isRequired
 };
 
-const mapStateToProps = ({ explore }) => ({
-  isLayerGroupAdded: dataset => !!explore.layers.find(l => l.dataset === dataset),
-  topicsTree: explore.topicsTree
+const mapStateToProps = state => ({
+  isLayerGroupAdded: dataset => !!state.explore.layers.find(l => l.dataset === dataset),
+  topicsTree: state.explore.topicsTree,
+  user: state.user
 });
 
 const mapDispatchToProps = dispatch => ({
   toggleLayerGroup: (dataset, addLayer) => dispatch(toggleLayerGroup(dataset, addLayer)),
-  toggleTooltip: (opened, opts) => { dispatch(toggleTooltip(opened, opts)); }
+  toggleTooltip: (opened, opts) => { dispatch(toggleTooltip(opened, opts)); },
+  addFavoriteDataset: (favorite) => { dispatch(addFavoriteDataset(favorite)); },
+  removeFavoriteDataset: (favorite) => { dispatch(removeFavoriteDataset(favorite)); }
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(DatasetWidget);